backend/requirements.txt
Django

djangorestframework

djangorestframework-simplejwt

django-cors-headers

Pillow
